fix: address review feedback for Range editor
